+++ src/components/atoms/FriendsList/FriendsListC/FriendsListC.stories.tsx
@@ -1,4 +1,5 @@
 import { Meta, StoryObj } from '@storybook/react'
+
 import FriendsListC from '.'
 
 const meta: Meta<typeof FriendsListC> = {
