+++ src/constants/tasteBoxImage.ts
@@ -1,17 +1,17 @@
-import { TasteBoxVariantType } from './apis/favor/TasteType'
+import { TasteBoxVariantType } from '../utils/apis/favor/TasteType'
 
 export const tasteBoxImage: Record<TasteBoxVariantType, string> = {
   LIP: '/images/makeup.png',
   EYE: '/images/eyeshadow.png',
   PERFUME: '/images/fragrance.png',
-  MOISTURE: '/images/candle.png', // ?
-  PLACE: '/images/candle.png', // ?
+  MOISTURE: '/images/candle.png',
+  PLACE: '/images/candle.png',
   TOP: '/images/clothes.png',
-  FAS_PRODUCT: '/images/keyring.png', // ?
-  DIG_PRODUCT: '/images/makeup.png', // ?
+  FAS_PRODUCT: '/images/keyring.png',
+  DIG_PRODUCT: '/images/keyring.png',
   BAG: '/images/bag.png',
   ACCESSORY: '/images/ring.png',
-  DISH: '/images/cup.png', // ?
+  DISH: '/images/cup.png',
   CUP: '/images/cup.png',
   EXERCISE: '/images/exercise.png',
 }
